tls::crypto::cipher

Trait MessageDecrypter

pub trait MessageDecrypter: Send + Sync {
    // Required method
    fn decrypt<'a>(
        &self,
        msg: InboundOpaqueMessage<'a>,
        seq: u64,
    ) -> Result<InboundPlainMessage<'a>, Error>;
}

A trait representing a message decrypter.

This trait provides a method to decrypt messages.

Required Methods

fn decrypt<'a>( &self, msg: InboundOpaqueMessage<'a>, seq: u64, ) -> Result<InboundPlainMessage<'a>, Error>

Decrypts the given message with the provided sequence number.

Arguments

msg - The inbound opaque message to decrypt.

seq - The sequence number.

Returns

A result containing the inbound plain message or an error.

Implementations

impl dyn MessageDecrypter

pub fn invalid() -> Box<dyn MessageDecrypter>

Returns an invalid message decrypter.

Returns

A boxed Invalid decrypter.