tls::crypto::cipher

Trait MessageEncrypter

pub trait MessageEncrypter: Sync + Send {
    // Required methods
    fn encrypt(
        &self,
        msg: OutboundPlainMessage<'_>,
        seq: u64,
    ) -> Result<OutboundOpaqueMessage, Error>;
    fn encrypted_payload_len(&self, payload_len: usize) -> usize;
}

A trait representing a message encrypter.

This trait provides methods to encrypt messages and to get the encrypted payload length.

Required Methods

fn encrypt( &self, msg: OutboundPlainMessage<'_>, seq: u64, ) -> Result<OutboundOpaqueMessage, Error>

Encrypts the given message with the provided sequence number.

Arguments

msg - The outbound plain message to encrypt.

seq - The sequence number.

Returns

A result containing the outbound opaque message or an error.

fn encrypted_payload_len(&self, payload_len: usize) -> usize

Returns the length of the encrypted payload.

Arguments

payload_len - The length of the payload.

Returns

The length of the encrypted payload.

Implementations

impl dyn MessageEncrypter

pub fn invalid() -> Box<dyn MessageEncrypter>

Returns an invalid message encrypter.

Returns

A boxed Invalid encrypter.