tls::crypto::hash

Trait Context

pub trait Context: Send + Sync {
    // Required methods
    fn finish(self: Box<Self>) -> Output;
    fn update(&mut self, data: &[u8]);
    fn fork_finish(&self) -> Output;
}

A trait representing a hash context.

This trait provides methods to finish the hash, update the hash with data, and fork the hash context.

Required Methods

fn finish(self: Box<Self>) -> Output

Finishes the hash and returns the result.

Returns

An Output containing the hash result.

fn update(&mut self, data: &[u8])

Updates the hash with the provided data.

Arguments

data - A slice of data to update the hash with.

fn fork_finish(&self) -> Output

Forks the hash context and finishes the forked context.

Returns

An Output containing the hash result of the forked context.