tls::crypto::hash

Trait Hash

pub trait Hash: Sync + Debug {
    // Required methods
    fn start(&self) -> Box<dyn Context>;
    fn hash(&self, data: &[u8]) -> Output;
    fn algorithm(&self) -> HashAlgorithm;
    fn output_length(&self) -> usize;
}

A trait representing a hash function.

This trait provides methods to start a new hash context, hash data, get the hash algorithm, and get the output length of the hash.

Required Methods

fn start(&self) -> Box<dyn Context>

Starts a new hash context.

Returns

A boxed Context for the hash.

fn hash(&self, data: &[u8]) -> Output

Hashes the provided data.

Arguments

data - A slice of data to hash.

Returns

An Output containing the hash result.

fn algorithm(&self) -> HashAlgorithm

Returns the hash algorithm.

Returns

The HashAlgorithm used.

fn output_length(&self) -> usize

Returns the output length of the hash.

Returns

The length of the hash output in bytes.