tls::crypto::hmac

Trait Key

pub trait Key: Sync {
    // Required methods
    fn sign(&self, data: &[&[u8]]) -> Tag;
    fn tag_len(&self) -> usize;
}

A trait representing a keyed HMAC instance.

This trait provides methods to sign data and to get the tag length.

Required Methods

fn sign(&self, data: &[&[u8]]) -> Tag

Signs the provided data.

Arguments

data - A slice of byte slices representing the data to sign.

Returns

A Tag containing the HMAC result.

fn tag_len(&self) -> usize

Returns the length of the tag.

Returns

The length of the tag in bytes.