tls::crypto::kx

Trait ActiveKx

pub trait ActiveKx: Send + Sync {
    // Required methods
    fn complete(self: Box<Self>, peer_pub_key: &[u8]) -> Result<Vec<u8>, Error>;
    fn group(&self) -> NamedCurve;
    fn pub_key(&self) -> &[u8] ;
}

A trait representing an active key exchange.

This trait provides methods to complete the key exchange, get the group, and get the public key.

Required Methods

fn complete(self: Box<Self>, peer_pub_key: &[u8]) -> Result<Vec<u8>, Error>

Completes the key exchange.

Arguments

peer_pub_key - A slice of bytes representing the peer’s public key.

Returns

A result containing a vector of bytes or an error.

fn group(&self) -> NamedCurve

Returns the key exchange group.

Returns

The NamedCurve representing the group.

fn pub_key(&self) -> &[u8]

Returns the public key.

Returns

A slice of bytes representing the public key.