tls::crypto::sign

Trait Signer

pub trait Signer {
    // Required methods
    fn sign(&self, message: &[u8]) -> Result<Vec<u8>, Error>;
    fn scheme(&self) -> SignatureScheme;
}

A trait representing a signer.

This trait provides methods to sign a message and to get the signature scheme.

Required Methods

fn sign(&self, message: &[u8]) -> Result<Vec<u8>, Error>

Signs the provided message.

Arguments

message - A slice of bytes representing the message to sign.

Returns

A result containing a vector of bytes or an error.

fn scheme(&self) -> SignatureScheme

Returns the signature scheme.

Returns

The SignatureScheme used.