Module tls::dynamic

Dynamically initialized, owning TLS

This module, built on top of the statik module, implements dynamically initialized TLS. This means that the first reference to a Key will run the dynamic initialization routine, and all accesses afterwards will reference the same contents.

As with "static TLS", this flavor of TLS also supports values with destructors, but has the same caveats as the static module.

Example

use std::cell::RefCell;

dynamic_tls!(static FOO: RefCell<Vec<uint>> = RefCell::new(vec![1, 2, 3]))

let vec = FOO.get().unwrap();
assert_eq!(vec.borrow().len(), 3);
vec.borrow_mut().push(4);
assert_eq!(vec.borrow()[3], 4);

Structs

Key
A dynamically initialized TLS key.

Ref
A reference to a Key.