Struct tls::dynamic::Key

pub struct Key<T: 'static> {
    // some fields omitted
}

A dynamically initialized TLS key.

This form of TLS key represents an owned value in TLS which is dynamically initialized upon the first reference to the value.

Example

use std::cell::RefCell;

dynamic_tls!(static FOO: RefCell<Vec<uint>> = RefCell::new(vec![1, 2, 3]))

let vec = FOO.get().unwrap();
assert_eq!(vec.borrow().len(), 3);
vec.borrow_mut().push(4);
assert_eq!(vec.borrow()[3], 4);

Methods

impl<T: 'static> Key<T>

fn get(&'static self) -> Option<Ref<T>>

Access this TLS variable, lazily initializing it if necessary.

The first time this function is called on each thread the TLS key will be initialized by having the specified init expression evaluated on the current thread.

This function can return None for the same reasons of static TLS returning None (destructors are running or may have run).