tls::futures

Struct Stream

pub struct Stream<'a, IO> {
    pub(crate) io: &'a mut IO,
    pub(crate) session: &'a mut Connection,
    eof: bool,
}

A structure representing a stream.

This structure contains a mutable reference to an I/O object and a mutable reference to a TLS connection session. It also keeps track of whether the end-of-file (EOF) has been reached.

Fields
io: &'a mut IO

A mutable reference to the I/O object.

session: &'a mut Connection

A mutable reference to the TLS connection session.

eof: bool

A boolean indicating whether the end-of-file (EOF) has been reached.

Implementations

impl<'a, IO> Stream<'a, IO>
where IO: AsyncRead + AsyncWrite + Unpin,

fn new(session: &'a mut Connection, io: &'a mut IO) -> Self

Creates a new Stream instance.

Arguments

session - A mutable reference to a Connection.

io - A mutable reference to an I/O object.

Returns

A new Stream instance.

fn set_eof(self, eof: bool) -> Self

Sets the end-of-file (EOF) flag.

Arguments

eof - A boolean indicating whether the EOF has been reached.

Returns

The Stream instance with the updated EOF flag.

fn as_mut_pin(&mut self) -> Pin<&mut Self>

Converts the Stream instance to a pinned mutable reference.

Returns

A pinned mutable reference to the Stream instance.

pub fn read_io(&mut self, cx: &mut Context<'_>) -> Poll<Result<usize>>

Reads data from the I/O object into the TLS session.

Arguments

cx - A mutable reference to the task context.

Returns

A Poll indicating the result of the read operation.

pub fn write_io(&mut self, cx: &mut Context<'_>) -> Poll<Result<usize>>

Writes data from the TLS session to the I/O object.

Arguments

cx - A mutable reference to the task context.

Returns

A Poll indicating the result of the write operation.

pub fn handshake( &mut self, cx: &mut Context<'_>, ) -> Poll<Result<(usize, usize)>>

Performs the TLS handshake.

Arguments

cx - A mutable reference to the task context.

Returns

A Poll indicating the result of the handshake operation.

fn poll_fill_buf(self, cx: &mut Context<'_>) -> Poll<Result<&'a [u8]>>

Fills the buffer with data from the TLS session.

Arguments

cx - A mutable reference to the task context.

Returns

A Poll indicating the result of the buffer fill operation.