tls::futures

Trait Session

trait Session {
    type IO;
    type Session;

    // Required methods
    fn get_mut(&mut self) -> (&mut TlsState, &mut Self::IO, &mut Self::Session);
    fn into_io(self) -> Self::IO;
}

A trait representing a session.

This trait provides methods to get mutable references to the state, I/O object, and session, and to convert the session into the I/O object.

Required Associated Types

type IO

type Session

Required Methods

fn get_mut(&mut self) -> (&mut TlsState, &mut Self::IO, &mut Self::Session)

Gets mutable references to the state, I/O object, and session.

Returns

A tuple containing mutable references to the state, I/O object, and session.

fn into_io(self) -> Self::IO

Converts the session into the I/O object.

Returns

The I/O object.