tls::message::base

Struct PayloadU8

pub struct PayloadU8(pub Vec<u8>);

Represents a payload with a length prefix of 1 byte.

Tuple Fields
0: Vec<u8>
Implementations

impl PayloadU8

pub fn new(body: Vec<u8>) -> Self

Creates a new payload with the given body.

Arguments

body - A vector of bytes.

Returns

A new payload with a length prefix of 1 byte.

pub fn new_empty() -> Self

Creates a new empty payload.

Returns

An empty payload with a length prefix of 1 byte.

fn encode_slice(slice: &[u8], bytes: &mut Vec<u8>)

Encodes a slice of bytes with a length prefix of 1 byte.

Arguments

slice - A slice of bytes to encode.

bytes - A mutable reference to a byte vector where the encoded slice will be stored.