tls::message

Module enums

Enums

CipherSuite

Represents the cipher suite in TLS.

CompressionMethod

Represents the compression method in TLS.

ContentType

Represents the content type in TLS.

ECCurveType 🔒

The ECCurveType TLS protocol enum. Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.

ECPointFormat

The ECPointFormat TLS protocol enum. Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.

ExtensionType

Represents the extension type in TLS.

HashAlgorithm

Represents the hash algorithm in TLS.

KeyExchangeAlgorithm

Represents the key exchange algorithm in TLS.

NamedCurve

Represents the named curve in TLS.

ProtocolVersion

Represents the protocol version in TLS.

ServerNameType 🔒

The ServerNameType TLS protocol enum. Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.

SignatureAlgorithm

Represents the signature algorithm in TLS.

SignatureScheme

The SignatureScheme TLS protocol enum. Values in this enum are taken from the various RFCs covering TLS, and are listed by IANA. The Unknown item is used when processing unrecognised ordinals.