tls::message

Module hs

Structs

CertificateChain

Represents a chain of certificates in the TLS protocol.

ClientECDHParams 🔒

Represents the parameters for ECDH key exchange in the ClientHello message.

ClientHelloPayload

Represents the payload of a ClientHello message in the TLS protocol.

Random

Represents a random value used in the TLS protocol.

ServerECDHParams

Represents the parameters for ECDH key exchange in the TLS protocol.

ServerHelloPayload

Represents the payload of a ServerHello message in the TLS protocol.

ServerKeyExchange

Represents the ServerKeyExchange message in the TLS protocol.

ServerName

Represents a server name in the TLS protocol.

SessionID

Represents a session ID in the TLS protocol.

SignatureAndHashAlgorithm

Represents a signature and hash algorithm in the TLS protocol.

UnknownExtension

Represents an unknown extension in the TLS protocol.

Enums

ClientExtension

Represents a client extension in the TLS protocol.

HandshakePayload

Enum representing different types of handshake payloads in the TLS protocol.

HandshakeType

Enum representing different types of handshake messages in the TLS protocol.

ServerExtension

Represents a server extension in the TLS protocol.

ServerKeyExchangePayload

Represents the payload of a ServerKeyExchange message in the TLS protocol.

ServerNamePayload

Represents the payload of a server name in the TLS protocol.

Constants

MAX_CERTIFICATE_SIZE_LIMIT

The maximum size limit for a certificate.