tls::message::hs

Struct ClientHelloPayload

pub struct ClientHelloPayload {
    pub client_version: ProtocolVersion,
    pub random: Random,
    pub session: SessionID,
    pub cipher_suites: Vec<CipherSuite>,
    pub compression_methods: Vec<CompressionMethod>,
    pub extensions: Vec<ClientExtension>,
}

Represents the payload of a ClientHello message in the TLS protocol.

Fields
client_version: ProtocolVersion

The protocol version used by the client.

random: Random

The random value generated by the client.

session: SessionID

The session ID provided by the client.

cipher_suites: Vec<CipherSuite>

The list of cipher suites supported by the client.

compression_methods: Vec<CompressionMethod>

The list of compression methods supported by the client.

extensions: Vec<ClientExtension>

The list of extensions provided by the client.

Implementations

impl ClientHelloPayload

pub(crate) fn find_extension( &self, typ: ExtensionType, ) -> Option<&ClientExtension>

Finds a specific extension in the list of client extensions.

Arguments

typ - The type of the extension to find.

Returns

Option<&ClientExtension> - A reference to the found extension or None if not found.

pub(crate) fn signature_algorithm(&self) -> Option<&[SignatureScheme]>

Returns the list of signature algorithms supported by the client.

Returns

Option<&[SignatureScheme]> - A reference to the list of signature schemes or None if not found.

pub(crate) fn named_groups(&self) -> Option<&[NamedCurve]>

Returns the list of named groups supported by the client.

Returns

Option<&[NamedCurve]> - A reference to the list of named curves or None if not found.

pub(crate) fn ecpoints(&self) -> Option<&[ECPointFormat]>

Returns the list of EC point formats supported by the client.

Returns

Option<&[ECPointFormat]> - A reference to the list of EC point formats or None if not found.

pub(crate) fn supports_ems(&self) -> bool

Checks if the client supports the Extended Master Secret extension.

Returns

bool - true if the client supports the Extended Master Secret extension, false otherwise.

pub(crate) fn sni_extension(&self) -> Option<&[ServerName]>

Returns the Server Name Indication (SNI) extension provided by the client.

Returns

Option<&[ServerName]> - A reference to the list of server names or None if not found.