Module tls::os

OS-based thread local storage

This module provides an implementation of OS-based thread local storage, using the native OS-provided facilities (think TlsAlloc or pthread_setspecific). The interface of this differs from the other types of thread-local-storage provided in this crate in that OS-based TLS can only get/set pointers,

This modules also provides two flavors of TLS. One is intended for static initialization, and does not contain a Drop implementation to deallocate the OS-TLS key. The other is a type which does implement Drop and hence has a safe interface.

Usage

This module should likely not be used directly unless other primitives are being built on. types such as tls::scoped::Key are likely much more useful in practice than this OS-based version which likely requires unsafe code to interoperate with.

Example

Using a dynamically allocated TLS key. Note that this key can be shared among many threads via an Arc.

use tls::os::Key;

let key = Key::new(None);
assert!(key.get().is_null());
key.set(1 as *mut u8);
assert!(!key.get().is_null());

drop(key); // deallocate this TLS slot.

Sometimes a statically allocated key is either required or easier to work with, however.

use tls::os::{StaticKey, INIT};

static KEY: StaticKey = INIT;

unsafe {
    assert!(KEY.get().is_null());
    KEY.set(1 as *mut u8);
}

Structs

Key
A type for a safely managed OS-based TLS slot.

StaticKey
A type for TLS keys that are statically allocated.

StaticKeyInner
Inner contents of StaticKey, created by the INIT_INNER constant.

Constants

INIT
Constant initialization value for static TLS keys.

INIT_INNER
Constant initialization value for the inner part of static TLS keys.