Struct tls::os::Key

pub struct Key {
    // some fields omitted
}

A type for a safely managed OS-based TLS slot.

This type allocates an OS TLS key when it is initialized and will deallocate the key when it falls out of scope. When compared with StaticKey, this type is entirely safe to use.

Implementations will likely, however, contain unsafe code as this type only operates on *mut u8, an unsafe pointer.

Example

use tls::os::Key;

let key = Key::new(None);
assert!(key.get().is_null());
key.set(1 as *mut u8);
assert!(!key.get().is_null());

drop(key); // deallocate this TLS slot.

Methods

impl Key

fn new(dtor: Option<unsafe extern "C" fn(*mut u8)>) -> Key

Create a new managed OS TLS key.

This key will be deallocated when the key falls out of scope.

The argument provided is an optionally-specified destructor for the value of this TLS key. When a thread exits and the value for this key is non-null the destructor will be invoked. The TLS value will be reset to null before the destructor is invoked.

Note that the destructor will not be run when the Key goes out of scope.

fn get(&self) -> *mut u8

See StaticKey::get

fn set(&self, val: *mut u8)

See StaticKey::set