tls

Module pki_types

Module for handling PKI types in the TLS protocol.

Modules

alg_id

The PKIX AlgorithmIdentifier type, and common values.

pem

Low-level PEM decoding APIs.

Structs

AddrParseError

Failure to parse an IP address

AlgorithmIdentifier

A DER encoding of the PKIX AlgorithmIdentifier type:

CertificateDer

A DER-encoded X.509 certificate; as specified in RFC 5280

CertificateRevocationListDer

A Certificate Revocation List; as specified in RFC 5280

CertificateSigningRequestDer

A Certificate Signing Request; as specified in RFC 2986

Der

DER-encoded data, either owned or borrowed

DnsName

A type which encapsulates a string (borrowed or owned) that is a syntactically valid DNS name.

EchConfigListBytes

A TLS-encoded Encrypted Client Hello (ECH) configuration list (ECHConfigList); as specified in draft-ietf-tls-esni-18 4

InvalidDnsNameError

The provided input could not be parsed because it is not a syntactically-valid DNS Name.

InvalidSignature

A detail-less error when a signature is not valid.

Ipv4Addr

no_std implementation of std::net::Ipv4Addr.

Ipv6Addr

no_std implementation of std::net::Ipv6Addr.

PrivatePkcs1KeyDer

A DER-encoded plaintext RSA private key; as specified in PKCS#1/RFC 3447

PrivatePkcs8KeyDer

A DER-encoded plaintext private key; as specified in PKCS#8/RFC 5958

PrivateSec1KeyDer

A Sec1-encoded plaintext private key; as specified in RFC 5915

SubjectPublicKeyInfoDer

A DER-encoded SubjectPublicKeyInfo (SPKI), as specified in RFC 5280.

TrustAnchor

A trust anchor (a.k.a. root CA)

UnixTime

A timestamp, tracking the number of non-leap seconds since the Unix epoch.

Enums

IpAddr

no_std implementation of std::net::IpAddr.

PrivateKeyDer

A DER-encoded X.509 private key, in one of several formats

ServerName

Encodes ways a client can know the expected name of the server.

Traits

SignatureVerificationAlgorithm

An abstract signature verification algorithm.

Type Aliases

SubjectPublicKeyInfoDeprecated

A DER-encoded SubjectPublicKeyInfo (SPKI), as specified in RFC 5280.