Module tls::scoped

Scoped thread-local storage

This module provides the ability to generate scoped thread-local variables. In this sense, scoped indicates that TLS actually stores a reference to a value, and this reference is only placed in TLS for a scoped amount of time.

There are no restrictions on what types can be placed into a scoped TLS variable, but all scoped variables are initialized to the equivalent of null. Scoped TLS is useful when a value is present in TLS for a known period of time and it is not required for TLS to take ownership of the contents.

Example

scoped_tls!(static FOO: uint)

// Initially each scoped TLS slot is empty.
FOO.with(|slot| {
    assert_eq!(slot, None);
});

// When inserting a value into TLS, the value is only in place for the
// duration of the closure specified.
FOO.set(&1, || {
    FOO.with(|slot| {
        assert_eq!(slot.map(|x| *x), Some(1));
    });
});

Structs

Key
Type representing a TLS key corresponding to a reference to the type parameter T.