Struct tls::scoped::Key

pub struct Key<T> {
    // some fields omitted
}

Type representing a TLS key corresponding to a reference to the type parameter T.

Keys are statically allocated and can contain a reference to an instance of type T scoped to a particular lifetime. Keys provides two methods, set and with, both of which currently use closures to control the scope of their contents.

Methods

impl<T> Key<T>

fn set<R>(&'static self, t: &T, cb: || -> R) -> R

Insert a value into this scoped TLS slot for a duration of a closure.

While cb is running, the value t will be returned by get unless this function is called recursively inside of cb.

Upon return, this function will restore the previous TLS value, if any was available.

Example

scoped_tls!(static FOO: uint)

FOO.set(&100, || {
    let val = FOO.with(|v| *v.unwrap());
    assert_eq!(val, 100);

    // set can be called recursively
    FOO.set(&101, || {
        // ...
    });

    // Recursive calls restore the previous value.
    let val = FOO.with(|v| *v.unwrap());
    assert_eq!(val, 100);
});

fn with<R>(&'static self, cb: |Option<&T>| -> R) -> R

Get a value out of this scoped TLS variable.

This function takes a closure which receives the value of this TLS variable, if any is available. If this variable has not yet been set, then None is yielded.

Example

scoped_tls!(static FOO: uint)

FOO.with(|slot| {
    // work with `slot`
});