Module tls::statik

Statically initialized, owning TLS

This module implements a "flavor" of TLS where the TLS slot owns the data that it contains. All contents, however, must be statically initialized. Any value which does not contain references is permitted inside a key of this type. This means that owning TLS keys also allow for content with destructors.

The destructor for TLS values will run when the thread exits. There are some nuances about destructors, however:

A TLS key cannot be accessed while its destructor is running.

A TLS key may not be accessible after its destructor has run.

Repeately setting TLS keys during destruction may cause memory leaks.

A panic! in a TLS destructor will result in a process abort.

TLS destructors may not be run when the application exits (the entire process is exiting anyway).

It is generally recommended to avoid TLS from destructors themselves, and if required only doing so in a deterministic, non-cyclic fashion.

This form of TLS will also attempt to select the "fastest" implementation of TLS available for the target platform.

Example

use std::cell::UnsafeCell;

tls!(static FOO: UnsafeCell<uint> = UnsafeCell { value: 1 });

unsafe {
    let f = FOO.get().unwrap();
    assert_eq!(*f.get(), 1);
    *f.get() = 2;

    // each thread starts out with the initial value of 1
    spawn(move || {
        let f = FOO.get().unwrap();
        assert_eq!(*f.get(), 1);
        *f.get() = 3;
    });

    // we retain our original value of 2 despite the child thread
    assert_eq!(*FOO.get().unwrap().get(), 2);
}

Structs

Key
A TLS key which owns its contents.

Ref
A structure representing a reference to a TLS value.