Struct tls::statik::Key

pub struct Key<T> {
    // some fields omitted
}

A TLS key which owns its contents.

This TLS key uses the fastest possible TLS implementation available to it for the target platform. It is instantiated with the tls! macro and the primary method is the get method.

The get method returns an object which represents a shared reference to the TLS value for the current thread. The object cannot be sent across tasks.

Initialization and Destruction

Currently initialization and interior mutability must be done through UnsafeCell, but hopefully this will change!

Values stored into TLS support destructors, and their destructors will be run when a thread exits.

Example

use std::cell::UnsafeCell;

tls!(static FOO: UnsafeCell<uint> = UnsafeCell { value: 1 });

unsafe {
    let f = FOO.get().unwrap();
    assert_eq!(*f.get(), 1);
    *f.get() = 2;

    // each thread starts out with the initial value of 1
    spawn(move || {
        let f = FOO.get().unwrap();
        assert_eq!(*f.get(), 1);
        *f.get() = 3;
    });

    // we retain our original value of 2 despite the child thread
    assert_eq!(*FOO.get().unwrap().get(), 2);
}

Methods

impl<T: 'static> Key<T>

fn get(&'static self) -> Option<Ref<T>>

Acquire a reference to the value in this TLS key.

This may lazily initializes an OS-based TLS key. This method in general is quite cheap to call, however.

This function will return None if the TLS value is currently being destroyed, and it may also return None after the key has been destroyed.