Crate tls

Rust bindings for libressl's libtls For the authoritative source on the inner workings of libtls check the manpage.

Client

use std::io::Write;
use std::net::TcpStream;
tls::init();
let tcp = TcpStream::connect("google.com:443").unwrap();
let mut client = tls::new_client()
    .from_socket(&tcp, "google.com")
    .unwrap();
client.write("GET / HTTP/1.1\n\n".as_bytes()).unwrap();

Server

The library does not handle TCP listening and binding, you need to handle the TCP server accept() and then call TlsServer::accept

use std::net::TcpListener;
tls::init();
let srv = TcpListener::bind("127.0.0.1:0").unwrap();
let addr = srv.local_addr().unwrap();
let mut tls_srv = tls::new_server()
    .key_file("tests/private_key.key")
    .cert_file("tests/certificate.crt")
    .bind().unwrap();
// Accept TCP connection, and then start TLS over it
let tcp_conn = srv.incoming().next().unwrap().unwrap();
let mut tls_conn = tls_srv.accept(&tcp_conn).unwrap();

Certificate Verification

By default libtls will verify certificates using the system certificate store (usually defined as /etc/ssl/cert.pem). In some Linux flavours and in Windows this file does not exist and you will need to use one of the appropriate methods to load the correct certificates for your system - check the Builder classes for the ca methods.

Connection Lifetime

The from_socket() and accept() methods build new connections on top of RawFd/RawSocket. Destroying the TlsStream object WILL NOT cause the underlying sockets to be closed, you need to close them yourself.

Structs

ClientBuilder

ServerBuilder

TlsError

TlsServer
TLS Server, used to start TLS session over existing sockets.

TlsStream

Functions

init
Initialize libtls - make sure to call this before using the API Returns false if libtls failed to initialise.

new_client
Create a new TLS client

new_server
Create a new TLS server

Type Definitions

TlsResult
result type for TLS operations